+++ modules/apps/oauth2-provider/oauth2-provider-web/src/main/java/com/liferay/oauth2/provider/web/internal/display/context/AssignScopesModel.java
@@ -310,6 +310,7 @@ public class AssignScopesModel {
 			if (!Validator.isBlank(scopeAlias)) {
 				combinedAuthorizationModelsRelations.put(
 					authorizationModel, authorizationModelRelations);
+
 				continue;
 			}
