+++ modules/apps/oauth2-provider/oauth2-provider-web/src/main/java/com/liferay/oauth2/provider/web/internal/display/context/OAuth2AdminPortletDisplayContext.java
@@ -14,33 +14,19 @@
 
 package com.liferay.oauth2.provider.web.internal.display.context;
 
-import com.liferay.document.library.kernel.service.DLAppLocalServiceUtil;
-import com.liferay.document.library.kernel.util.DLUtil;
 import com.liferay.oauth2.provider.configuration.OAuth2ProviderConfiguration;
 import com.liferay.oauth2.provider.constants.GrantType;
-import com.liferay.oauth2.provider.constants.OAuth2ProviderActionKeys;
-import com.liferay.oauth2.provider.constants.OAuth2ProviderConstants;
 import com.liferay.oauth2.provider.model.OAuth2Application;
 import com.liferay.oauth2.provider.model.OAuth2ApplicationScopeAliases;
 import com.liferay.oauth2.provider.service.OAuth2ApplicationScopeAliasesLocalServiceUtil;
 import com.liferay.oauth2.provider.service.OAuth2ApplicationService;
 import com.liferay.oauth2.provider.service.OAuth2AuthorizationServiceUtil;
-import com.liferay.oauth2.provider.web.internal.constants.OAuth2AdminActionKeys;
-import com.liferay.oauth2.provider.web.internal.constants.OAuth2ProviderPortletKeys;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.io.BigEndianCodec;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.security.SecureRandomUtil;
-import com.liferay.portal.kernel.security.permission.ActionKeys;
-import com.liferay.portal.kernel.security.permission.PermissionChecker;
-import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
-import com.liferay.portal.kernel.service.permission.PortletPermissionUtil;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.StringUtil;
 
 import java.util.ArrayList;
@@ -50,13 +36,13 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import javax.portlet.PortletPreferences;
-
-import javax.servlet.http.HttpServletRequest;
+import javax.portlet.PortletRequest;
 
 /**
  * @author Tomas Polesovsky
  */
-public class OAuth2AdminPortletDisplayContext {
+public class OAuth2AdminPortletDisplayContext
+	extends BaseOAuth2PortletDisplayContext {
 
 	public static String generateRandomSecret() {
 		int size = 16;
@@ -81,35 +67,15 @@ public class OAuth2AdminPortletDisplayContext {
 	}
 
 	public OAuth2AdminPortletDisplayContext(
+		PortletRequest portletRequest,
 		OAuth2ApplicationService oAuth2ApplicationService,
 		OAuth2ProviderConfiguration oAuth2ProviderConfiguration,
 		ThemeDisplay themeDisplay) {
 
-		_oAuth2ApplicationService = oAuth2ApplicationService;
+		super.oAuth2ApplicationService = oAuth2ApplicationService;
 		_oAuth2ProviderConfiguration = oAuth2ProviderConfiguration;
-		_themeDisplay = themeDisplay;
-	}
-
-	public OAuth2Application getApplication(HttpServletRequest request)
-		throws PortalException {
-
-		if (_oAuth2Application != null) {
-			return _oAuth2Application;
-		}
-
-		long oAuth2ApplicationId = ParamUtil.getLong(
-			request, "oAuth2ApplicationId", 0);
-
-		if (oAuth2ApplicationId > 0) {
-			_oAuth2Application = _oAuth2ApplicationService.getOAuth2Application(
-				oAuth2ApplicationId);
-		}
-
-		return _oAuth2Application;
-	}
-
-	public String getDefaultIconURL() {
-		return _themeDisplay.getPathThemeImages() + "/common/portlet.png";
+		super.portletRequest = portletRequest;
+		super.themeDisplay = themeDisplay;
 	}
 
 	public int getOAuth2AuthorizationsCount(OAuth2Application oAuth2Application)
@@ -187,109 +153,9 @@ public class OAuth2AdminPortletDisplayContext {
 		return scopeAliasesList.size();
 	}
 
-	public String getThumbnailURL(OAuth2Application oAuth2Application)
-		throws Exception {
-
-		if (oAuth2Application.getIconFileEntryId() <= 0) {
-			return getDefaultIconURL();
-		}
-
-		FileEntry fileEntry = DLAppLocalServiceUtil.getFileEntry(
-			oAuth2Application.getIconFileEntryId());
-
-		return DLUtil.getThumbnailSrc(fileEntry, _themeDisplay);
-	}
-
-	public boolean hasAddApplicationPermission() {
-		PermissionChecker permissionChecker =
-			PermissionThreadLocal.getPermissionChecker();
-
-		if (permissionChecker.hasPermission(
-				0, OAuth2ProviderConstants.RESOURCE_NAME,
-				OAuth2ProviderConstants.RESOURCE_NAME,
-				OAuth2ProviderActionKeys.ACTION_ADD_APPLICATION)) {
-
-			return true;
-		}
-
-		return false;
-	}
-
-	public boolean hasDeletePermission(OAuth2Application oAuth2Application) {
-		return hasPermission(oAuth2Application, ActionKeys.DELETE);
-	}
-
-	public boolean hasPermission(
-		OAuth2Application oAuth2Application, String actionId) {
-
-		PermissionChecker permissionChecker =
-			PermissionThreadLocal.getPermissionChecker();
-
-		if (permissionChecker.hasOwnerPermission(
-				oAuth2Application.getCompanyId(),
-				OAuth2Application.class.getName(),
-				oAuth2Application.getOAuth2ApplicationId(),
-				oAuth2Application.getUserId(), actionId)) {
-
-			return true;
-		}
-
-		if (permissionChecker.hasPermission(
-				0, OAuth2Application.class.getName(),
-				oAuth2Application.getOAuth2ApplicationId(), actionId)) {
-
-			return true;
-		}
-
-		return false;
-	}
-
-	public boolean hasPermissionsPermission(
-		OAuth2Application oAuth2Application) {
-
-		return hasPermission(oAuth2Application, ActionKeys.PERMISSIONS);
-	}
-
-	public boolean hasRevokeTokenPermission(
-		OAuth2Application oAuth2Application) {
-
-		return hasPermission(
-			oAuth2Application, OAuth2ProviderActionKeys.ACTION_REVOKE_TOKEN);
-	}
-
-	public boolean hasUpdatePermission(OAuth2Application oAuth2Application) {
-		return hasPermission(oAuth2Application, ActionKeys.UPDATE);
-	}
-
-	public boolean hasViewGrantedAuthorizationsPermission() {
-		PermissionChecker permissionChecker =
-			PermissionThreadLocal.getPermissionChecker();
-
-		try {
-			return PortletPermissionUtil.contains(
-				permissionChecker, OAuth2ProviderPortletKeys.OAUTH2_ADMIN,
-				OAuth2AdminActionKeys.VIEW_GRANTED_AUTHORIZATIONS);
-		}
-		catch (PortalException pe) {
-			_log.error(pe);
-
-			return false;
-		}
-	}
-
-	public boolean hasViewPermission(OAuth2Application oAuth2Application) {
-		return hasPermission(oAuth2Application, ActionKeys.VIEW);
-	}
-
-	private static Log _log = LogFactoryUtil.getLog(
-		OAuth2AdminPortletDisplayContext.class);
-
 	private static final Pattern _baseIdPattern = Pattern.compile(
 		"(.{8})(.{4})(.{4})(.{4})(.*)");
 
-	private OAuth2Application _oAuth2Application;
-	private final OAuth2ApplicationService _oAuth2ApplicationService;
 	private final OAuth2ProviderConfiguration _oAuth2ProviderConfiguration;
-	private final ThemeDisplay _themeDisplay;
 
 }
