+++ modules/apps/oauth2-provider/oauth2-provider-web/src/main/resources/content/Language.properties
@@ -87,3 +87,8 @@ privacy-policy=Privacy Policy
 remove-access=Remove Access
 are-you-sure-you-want-to-remove-access-for-the-selected-entries=Are you sure you want to remove access for the selected entries?
 x-will-no-longer-have-access-to-your-account-removed-access-cannot-be-recovered={0} will no longer have access to your account. Removed access cannot be recovered.
+choose-one-of-the-following-global-scopes-that-include-this-resource-scope=Choose one of the following global scopes that include this resource scope...
+resource-scopes=Resource scopes
+global-scopes=Global scopes
+can-be-assigned-via-a-global-scope=CAN be assigned via a global scope
+must-be-assigned-via-a-global-scope=MUST be assigned via a global scope
